Implement 4_1

+++ include/aoc15/d04.h
@@ -3,6 +3,6 @@
 
 #include <string>
 
-int miningNumber(const std::string &key, int leadingZeroes);
+int miningNumber(const std::string &key, size_t leadingZeroes);
 
 #endif // AOC15_D04_H
